+++ NetFabric.Hyperlinq/Conversion/ToList/ToList.ReadOnlyList.cs
@@ -38,7 +38,7 @@ namespace NetFabric.Hyperlinq
                 return list;
         }
 
-        static List<TSource> ToList<TEnumerable, TSource>(this TEnumerable source, Func<TSource, long, bool> predicate, int skipCount, int takeCount)
+        static List<TSource> ToList<TEnumerable, TSource>(this TEnumerable source, Func<TSource, int, bool> predicate, int skipCount, int takeCount)
             where TEnumerable : IReadOnlyList<TSource>
         {
                 var list = new List<TSource>();
