+++ NetFabric.Hyperlinq/Element/First/First.ReadOnlyCollection.cs
@@ -66,7 +66,7 @@ namespace NetFabric.Hyperlinq
             return (ElementResult.Empty, default);
         }    
 
-        public static (int Index, TSource Value) TryFirst<TEnumerable, TEnumerator, TSource>(this TEnumerable source, Func<TSource, long, bool> predicate) 
+        public static (int Index, TSource Value) TryFirst<TEnumerable, TEnumerator, TSource>(this TEnumerable source, Func<TSource, int, bool> predicate) 
             where TEnumerable : IReadOnlyCollection<TSource>
             where TEnumerator : IEnumerator<TSource>
         {
