+++ NetFabric.Hyperlinq/Element/Single/Single.Array.cs
@@ -15,7 +15,7 @@ namespace NetFabric.Hyperlinq
             return ref source[0];
         }
 
-        public static ref TSource Single<TSource>(this TSource[] source, Func<TSource, long, bool> predicate)
+        public static ref TSource Single<TSource>(this TSource[] source, Func<TSource, int, bool> predicate)
         {
             if (predicate is null) ThrowHelper.ThrowArgumentNullException(nameof(predicate));
 
@@ -63,7 +63,7 @@ namespace NetFabric.Hyperlinq
             return ref source[0];
         }
 
-        public static ref TSource Single<TSource>(this TSource[] source, Func<TSource, long, bool> predicate, out int index)
+        public static ref TSource Single<TSource>(this TSource[] source, Func<TSource, int, bool> predicate, out int index)
         {
             if (predicate is null) ThrowHelper.ThrowArgumentNullException(nameof(predicate));
 
@@ -120,7 +120,7 @@ namespace NetFabric.Hyperlinq
             return ref Default<TSource>.Value;
         }
 
-        public static ref readonly TSource SingleOrDefault<TSource>(this TSource[] source, Func<TSource, long, bool> predicate)
+        public static ref readonly TSource SingleOrDefault<TSource>(this TSource[] source, Func<TSource, int, bool> predicate)
         {
             if (predicate is null) ThrowHelper.ThrowArgumentNullException(nameof(predicate));
 
@@ -143,7 +143,7 @@ namespace NetFabric.Hyperlinq
             return ref Default<TSource>.Value;
         }
 
-        public static ref readonly TSource SingleOrDefault<TSource>(this TSource[] source, Func<TSource, long, bool> predicate, out int index)
+        public static ref readonly TSource SingleOrDefault<TSource>(this TSource[] source, Func<TSource, int, bool> predicate, out int index)
         {
             if (predicate is null) ThrowHelper.ThrowArgumentNullException(nameof(predicate));
 
@@ -179,11 +179,11 @@ namespace NetFabric.Hyperlinq
         }
 
         [MethodImpl(MethodImplOptions.AggressiveInlining)]
-        public static TSource? SingleOrNull<TSource>(this TSource[] source, Func<TSource, long, bool> predicate)
+        public static TSource? SingleOrNull<TSource>(this TSource[] source, Func<TSource, int, bool> predicate)
             where TSource : struct
             => SingleOrNull<TSource>(source, predicate, out var _);
 
-        public static TSource? SingleOrNull<TSource>(this TSource[] source, Func<TSource, long, bool> predicate, out int index)
+        public static TSource? SingleOrNull<TSource>(this TSource[] source, Func<TSource, int, bool> predicate, out int index)
             where TSource : struct
         {
             if (predicate is null) ThrowHelper.ThrowArgumentNullException(nameof(predicate));
