+++ NetFabric.Hyperlinq/Element/Single/Single.ReadOnlyCollection.cs
@@ -79,7 +79,7 @@ namespace NetFabric.Hyperlinq
             return (ElementResult.Empty, default);
         }
 
-        public static (int Index, TSource Value) TrySingle<TEnumerable, TEnumerator, TSource>(this TEnumerable source, Func<TSource, long, bool> predicate) 
+        public static (int Index, TSource Value) TrySingle<TEnumerable, TEnumerator, TSource>(this TEnumerable source, Func<TSource, int, bool> predicate) 
             where TEnumerable : IReadOnlyCollection<TSource>
             where TEnumerator : IEnumerator<TSource>
         {
