+++ NetFabric.Hyperlinq/Quantifier/Any/Any.ReadOnlyCollection.cs
@@ -29,7 +29,7 @@ namespace NetFabric.Hyperlinq
             return false;
         }
 
-        public static bool Any<TEnumerable, TEnumerator, TSource>(this TEnumerable source, Func<TSource, long, bool> predicate)
+        public static bool Any<TEnumerable, TEnumerator, TSource>(this TEnumerable source, Func<TSource, int, bool> predicate)
             where TEnumerable : IReadOnlyCollection<TSource>
             where TEnumerator : IEnumerator<TSource>
         {
