+++ NetFabric.Hyperlinq/Utils/ThrowHelper.cs
@@ -16,16 +16,16 @@ namespace NetFabric.Hyperlinq
         public static T ThrowArgumentOutOfRangeException<T>(string paramName)
             => throw new ArgumentOutOfRangeException(paramName);
 
-        public static void ThrowArgumentTooLargeException(long count)
+        public static void ThrowArgumentTooLargeException(int count)
             => throw new ArgumentOutOfRangeException($"Collection with count {count} exceeds allowed size.");
 
-        public static T ThrowArgumentTooLargeException<T>(long count)
+        public static T ThrowArgumentTooLargeException<T>(int count)
             => throw new ArgumentOutOfRangeException($"Collection with count {count} exceeds allowed size.");
 
-        public static void ThrowArgumentTooLargeException(string paramName, long count)
+        public static void ThrowArgumentTooLargeException(string paramName, int count)
             => throw new ArgumentOutOfRangeException(paramName, $"Collection with count {count} exceeds allowed size.");
 
-        public static T ThrowArgumentTooLargeException<T>(string paramName, long count)
+        public static T ThrowArgumentTooLargeException<T>(string paramName, int count)
             => throw new ArgumentOutOfRangeException(paramName, $"Collection with count {count} exceeds allowed size.");
 
         public static void ThrowIndexOutOfRangeException()
